Create PhoneNumberProgram
Printed out a phone number and manipulated the string based on the number of digits. This program is good to understand how to manipulate strings. Manipulation through only allowing a certain amount of characters to be entered, inserting certain text/characters into a string, and not allowing certain characters into a string.